Update SearchResultItem test
